+++ TESTS/test_AudioFileHelpers.cpp
@@ -53,11 +53,11 @@ TEST_CASE("Get length in samples of wav file.")
 }
 
 //===============================
-// 
+//
 TEST_CASE("Get sampleRate of wav file.")
 {
     auto inputFile = RelativeFilePath::getGoldenFileFromProjectRoot("GOLDEN_Somewhere_Mono_441k.wav");
-    int sampleRate = AudioFileHelpers::getFileSampleRate(inputFile);
+    int sampleRate = static_cast<int>(AudioFileHelpers::getFileSampleRate(inputFile));
 
     CHECK(sampleRate == 44100);
 }
